Fix removing of overflowed records (#55)
Groups with wildcard patterns must not include channels from previous groups.
At startup the service reads already stored channels from DB and then populates channels for each group. The groups are checked as they declared in config. If a matching group is found, the search is stopped. This way the order of groups in config is important. The first matching group is used.